LPD-34110 Move logic to class, no logical changes
