+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/check/JavaAssertEqualsCheck.java
@@ -9,6 +9,7 @@ import com.liferay.petra.string.CharPool;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.source.formatter.check.util.JavaSourceUtil;
+import com.liferay.source.formatter.check.util.SourceUtil;
 
 import java.util.List;
 import java.util.regex.Matcher;
@@ -45,7 +46,8 @@ public class JavaAssertEqualsCheck extends BaseFileCheck {
 
 			String actualParameter = parametersList.get(1);
 
-			String strippedQuotesActualParameter = stripQuotes(actualParameter);
+			String strippedQuotesActualParameter = SourceUtil.stripQuotes(
+				actualParameter);
 
 			if (!actualParameter.startsWith("expected") &&
 				!Validator.isDigit(actualParameter) &&
